Refactor source paths
