+++ package.json
@@ -4,8 +4,7 @@
   "description": "An image generator with 20+ metrics about your GitHub account such as activity, community, repositories, coding habits, website performances, music played, starred topics, etc. that you can put on your profile or elsewhere !",
   "main": "index.mjs",
   "scripts": {
-    "start": "node index.mjs",
-    "build": "node utils/build.mjs",
+    "start": "node source/app/web/index.mjs",
     "test": "npx jest",
     "upgrade": "npm install @actions/core@latest @actions/github@latest @octokit/graphql@latest @octokit/rest@latest axios@latest colors@latest compression@latest ejs@latest express@latest express-rate-limit@latest image-to-base64@latest memory-cache@latest prismjs@latest puppeteer@latest svgo@latest vue@latest vue-prism-component@latest jest@latest js-yaml@latest libxmljs@latest"
   },
