+++ source/app/metrics.mjs
@@ -1,15 +1,15 @@
 //Imports
-  import ejs from "ejs"
-  import SVGO from "svgo"
-  import imgb64 from "image-to-base64"
-  import axios from "axios"
-  import puppeteer from "puppeteer"
-  import url from "url"
-  import processes from "child_process"
   import fs from "fs/promises"
   import os from "os"
   import paths from "path"
   import util from "util"
+  import axios from "axios"
+  import url from "url"
+  import puppeteer from "puppeteer"
+  import processes from "child_process"
+  import ejs from "ejs"
+  import imgb64 from "image-to-base64"
+  import SVGO from "svgo"
 
 //Setup
   export default async function metrics({login, q, dflags = []}, {graphql, rest, plugins, conf, die = false, verify = false}, {Plugins, Templates}) {
