+++ source/plugins/music/index.mjs
