+++ source/templates/terminal/template.mjs
